Implements automatic sign in on sign up
Sends both a POST request to create a new user's profile and a POST
request to sign that user in when the sign up form is submitted. Sign
up success function is no longer necessary in the ui file, as all sign
ups are now also sign ins, which will have a more meaningful success
function in the future.